+++ botlink/src/main.rs
@@ -4,7 +4,7 @@ use log::info;
 use std::thread;
 const VERSION: &str = env!("CARGO_PKG_VERSION");
 
-use botlink::registry::Components;
+use botlink::registry::{create_redis_client, Components};
 
 #[tokio::main]
 async fn main() {
@@ -12,7 +12,11 @@ async fn main() {
     botlink::env::init();
     info!("🔢 {}", VERSION);
 
-    let components = Components::default();
+    let client = create_redis_client();
+
+    stream::init::create_consumer_group(&client);
+
+    let components = Components::new(client);
     let ws_opts = websocket::WSOpts::from(&components);
     let mco = components.move_computed_out.clone();
     let xmm = components.xadder.clone();