+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # NEXT: v0.9.1
 
 - gateway: fix deprecated time call (#426, #413)
-- Use XREADGROUP in gateway & botlink, upgrade their redis dep (#302, #310, #420, #423)
+- Use XREADGROUP in gateway & botlink, upgrade their redis dep (#302, #310, #420, #423, #427)
 - Simplify redis-streams crate (#421)
 
 # v0.9.0